Reduce code dup
