+++ src/client/datascience/jupyter/kernels/jupyterKernelSpec.ts
@@ -2,6 +2,11 @@
 // Licensed under the MIT License.
 'use strict';
 import { Kernel } from '@jupyterlab/services';
+import * as path from 'path';
+import { CancellationToken } from 'vscode';
+import { createPromiseFromCancellation } from '../../../common/cancellation';
+import { traceInfo } from '../../../common/logger';
+import { IFileSystem } from '../../../common/platform/types';
 import { PythonInterpreter } from '../../../interpreter/contracts';
 import { IJupyterKernelSpec } from '../../types';
 
@@ -23,4 +28,33 @@ export class JupyterKernelSpec implements IJupyterKernelSpec {
         this.display_name = specModel.display_name;
         this.metadata = specModel.metadata;
     }
+}
+
+export async function parseKernelSpecs(stdout: string, fs: IFileSystem, token?: CancellationToken) {
+    traceInfo('Parsing kernelspecs from jupyter');
+    // This should give us back a key value pair we can parse
+    const jsOut = JSON.parse(stdout.trim()) as { kernelspecs: Record<string, { resource_dir: string; spec: Omit<Kernel.ISpecModel, 'name'> }> };
+    const kernelSpecs = jsOut.kernelspecs;
+
+    const specs = await Promise.race([
+        Promise.all(
+            Object.keys(kernelSpecs).map(async kernelName => {
+                const specFile = path.join(kernelSpecs[kernelName].resource_dir, 'kernel.json');
+                const spec = kernelSpecs[kernelName].spec;
+                // Add the missing name property.
+                const model = {
+                    ...spec,
+                    name: kernelName
+                };
+                // Check if the spec file exists.
+                if (await fs.fileExists(specFile)) {
+                    return new JupyterKernelSpec(model as Kernel.ISpecModel, specFile);
+                } else {
+                    return;
+                }
+            })
+        ),
+        createPromiseFromCancellation({ cancelAction: 'resolve', defaultValue: [], token })
+    ]);
+    return specs.filter(item => !!item).map(item => item as JupyterKernelSpec);
 }
